Update original_game.py
When I ran this file alone with "./" on a windows 10 system with python 3.8, I was given file or directory not found error. After changing the value of ASSETS_DIR to "../flappy_bird_gym/assets", this problem went away and I was able to play the game. I suspect this may only be an issue with windows systems, as that is how we go back a directory in the command prompt.